fix classes

+++ libwdiv/include/compiler.hpp
@@ -182,7 +182,7 @@ private:
   void emitBytes(uint8 byte1, uint8 byte2);
   void emitReturn();
   void emitConstant(Value value);
-  int makeConstant(Value value);
+  uint8 makeConstant(Value value);
 
   int emitJump(uint8 instruction);
   void patchJump(int offset);
@@ -239,21 +239,21 @@ private:
   void emitGosubTo(int targetOffset);
   void patchJumpTo(int operandOffset, int targetOffset);
 
-  void handle_assignment(uint16 getOp, uint16 setOp, int arg, bool canAssign);
+  void handle_assignment(uint8 getOp, uint8 setOp, int arg, bool canAssign);
 
   void prefixIncrement(bool canAssign);
   void prefixDecrement(bool canAssign);
 
   // Variables
-  uint16 identifierConstant(Token &name);
+  uint8 identifierConstant(Token &name);
   void namedVariable(Token &name, bool canAssign);
-  void defineVariable(uint16 global);
+  void defineVariable(uint8 global);
   void declareVariable();
   void addLocal(Token &name);
   int resolveLocal(Token &name);
   void markInitialized();
 
-  uint16 argumentList();
+  uint8 argumentList();
 
   void compileFunction(Function *func, bool isProcess);
   void compileProcess(const std::string &name);
